backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.16
langchain-openai==0.1.3
langgraph==0.0.40
chromadb==0.4.24
pandas==2.2.1
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.0